+++ FamilyPets.Backend/src/FamilyPets.Domain/FullName.cs
@@ -1,6 +1,6 @@
 ﻿using CSharpFunctionalExtensions;
 
-namespace FamilyForPets.Domain.Volunteer
+namespace FamilyForPets.Domain
 {
     public class FullName : ValueObject
     {
