+++ FamilyPets.Backend/src/FamilyPets.Domain/VolunteerAgregate/PetValueObjects/PelageColor.cs
@@ -1,7 +1,7 @@
 ﻿using System.Drawing;
 using CSharpFunctionalExtensions;
 
-namespace FamilyForPets.Domain.Volunteer
+namespace FamilyForPets.Domain.VolunteerAgregate.PetValueObjects
 {
     public class PelageColor : ValueObject
     {
