+++ FamilyPets.Backend/src/FamilyPets.Domain/VolunteerAgregate/VolunteerValueObjects/SocialNetwork.cs
@@ -1,6 +1,6 @@
 ﻿using CSharpFunctionalExtensions;
 
-namespace FamilyForPets.Domain.Volunteer
+namespace FamilyForPets.Domain.VolunteerAgregate.VolunteerValueObjects
 {
     public class SocialNetwork : ValueObject
     {
