--- FamilyPets.Backend/src/FamilyPets.Infrastructure/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace FamilyPets.Infrastructure
-{
-    public class Class1
-    {
-
-    }
-}
