--- FamilyPets.Backend/src/FamilyPets.UseCases/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace FamilyPets.UseCases
-{
-    public class Class1
-    {
-
-    }
-}
